+++ packages/mini_program_tooling/lib/src/miniprogram_cli.dart
@@ -33,7 +33,7 @@ const List<String> _supportedPublishTargets = <String>[
   'firebase-hosting',
 ];
 
-const String _miniProgramToolingVersion = '0.3.40';
+const String _miniProgramToolingVersion = '0.3.42';
 
 const List<String> _capabilityIds = <String>[
   'publish.firebase_hosting',